Moves addition later so we don't cache info from the dynakube
